change name from case to project; fixing bugs on deletion of related objects

+++ src/main/java/pl/qceyco/app/client/ClientsAllController.java
@@ -1,11 +1,14 @@
 package pl.qceyco.app.client;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import pl.qceyco.app.client.additionalInfo.AdditionalInfoClientRepository;
+import pl.qceyco.app.project.ProjectRepository;
 
 import java.util.List;
 
@@ -14,12 +17,14 @@ import java.util.List;
 
 public class ClientsAllController {
 
+    private final ProjectRepository projectRepository;
     private final ClientsAllRepository clientsAllRepository;
     private final AdditionalInfoClientRepository additionalInfoClientRepository;
 
-    public ClientsAllController(ClientsAllRepository clientsAllRepository, AdditionalInfoClientRepository additionalInfoClientRepository) {
+    public ClientsAllController(ClientsAllRepository clientsAllRepository, AdditionalInfoClientRepository additionalInfoClientRepository, ProjectRepository projectRepository) {
         this.clientsAllRepository = clientsAllRepository;
         this.additionalInfoClientRepository = additionalInfoClientRepository;
+        this.projectRepository = projectRepository;
     }
 
     @ModelAttribute("clients")
@@ -40,7 +45,11 @@ public class ClientsAllController {
     }
 
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, Model model) {
+        if (projectRepository.findFirstByClientId(id) != null) {
+            model.addAttribute("deleteErrorProjectExists", "Cannot delete this client - delete related project first!");
+            return "clients/clientsList";
+        }
         Client clientToDelete = clientsAllRepository.findFirstById(id);
         clientsAllRepository.deleteById(id);
         if (clientToDelete.getAdditionalInfo() != null) {
